Move xpath tools to separate module
