+++ lib/versionista.js
@@ -4,6 +4,7 @@ const crypto = require('crypto');
 const request = require('request');
 const jsdom = require('jsdom');
 const Entities = require('html-entities').AllHtmlEntities;
+const {xpath, xpathArray, xpathNode} = require('./xpath');
 
 const MAX_SOCKETS = 6;
 const USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36';
@@ -515,42 +516,11 @@ function hash (text) {
   return crypto.createHash('sha256').update(text).digest('hex');
 }
 
-function xpath (node, expression) {
-  const document = node.nodeType === node.DOCUMENT_NODE ? node : node.ownerDocument;
-  const type = document.defaultView.XPathResult.ORDERED_NODE_ITERATOR_TYPE;
-  const iterator = document.evaluate(expression, node, null, type, null);
-  iterator.map = function (transform) {
-    let item;
-    let result = [];
-    while (item = iterator.iterateNext()) {
-      result.push(transform(item));
-    }
-    return result;
-  }
-  return iterator;
-}
-
-function xpathArray (node, expression) {
-  return xpath(node, expression).map(item => item);
-}
-
-function xpathNode (node, expression) {
-  const iterator = xpath(node, expression);
-  return iterator.iterateNext();
-}
-
 function joinUrlPaths (basePath, ...paths) {
   return paths.reduce((finalPath, urlPath) => {
     const delimiter = finalPath.endsWith('/') ? '' : '/';
     return finalPath + delimiter + urlPath;
   }, basePath);
-}
-
-function promisedInput (func) {
-  return function () {
-    return Promise.all(Array.from(arguments).map(Promise.resolve))
-      .then(resolvedArgs => func.apply(this, resolvedArgs))
-  }
 }
 
 function getPagingUrls (window) {
